INITIAL R&D CHANGES
The file mat.py has been created. We want to compare arrays between the coordinates that have E in it vs the ones without E. THis will help us choose which lines to plot and which lines to not plot. After this, we can incorporate the small proof of concept code into the overall simulation.py class.