Code Update
This technically works in making the simulation run a lot faster, however it seems that we may want to make it run even faster than this! We need to be able to run the simulation without rendering all lines. Perhaps we need to explore a way of keeping them as a fixed object on the screen without having to redraw with each rotate