Update 2
In this update, we have migrated test functions. We have made a separate file called 2vista.py in which we've made the first steps for animating a layer by layer GIF viewer instead of the static view. This is decent but we should edit it by SHOWING the gif rather than just saving it, allowing movement during animation. We should then also keep each previous plot so it's more overlayed