Configuration Update
This commit pushes updated code related to the configuration file(s) for use with the ender3. Temperature and other issues have been addressed. Furthermore, updated logic has been included with regards to setting a filename (previously it was static to the first file in the output directory). Now we scan the original configs file and see what it SHOULD be.