File Hanlding for Working Merge
Patch fix for config bundling has been pushed. This commit includes changes to file locations for cleanliness for the upcoming pushes. Config Parsing has been inregrated with the rest of the code, however we have some issues with the simulation and animation render tool as a result which needs fixing in the upcoming updates.