Method update
Centre of mass Method update. We can now accurately get CoM howver we need to consider cases when the object we pick is hollow. We may need to start working on a GUI or try to make a GUI tool or implement with prusa GUI to select where to pick and place with the STl prior to slicing