MAJOR UPDATE
Automated Implementation of the Stl Viewer with use inside the Tools/Vacuum_PnP method. Now we can choose where we want to use the tool. The only thing is this may need implementation/consideratoin with where to inject it into the file as that was still being done layer by layer. Now it generates where, but maybe the point needs finding roughly (i.e. find thenearest layer) to inject>