Initial Config Update
The configuration file will be changing form in the next update to allow for the user to have the option of setting their own centre. The user will be able to add in their centre coordinates for printing. If a coordinate for the centre is not supplied we will run a different command (without the centre argument included). The next update will use this, with further updates integrating this update to the actual PrusaSlicer class