Centre Positional Update
In this major update we have updated the centre positioning so that the STL can be printed at a given centre position. Logic has been included so if this is not provided it defaults to the centre of the bed. In the case of a centre, a new position will be calculated. Dynamic offset calculation is now done to the offsets from yesterday (for x y and z) yielding more accurate positioning. The AMPIConfigFile.ini has also been updated to stop killing the motors after every print