Bed Positional Updat
In this update we have succesfully implemented bed positoining calibration. We have worked out the offset from the bed if a position is given. While this is calibrated, we need to handle the event that we don't have a centre position provided. This needs to be done on the STLViewer end