Patch Fix and Update
Updated the vacuum logic so it now vibrates and separates the part from the bed before gripping with the vacuum tool. This will need automation so that we can separate the part from bed given the centre coordinates (of x and y, and z separte z value) while keeping x y and z of the other values.