Formatting Change
This commit pushes code which separates the vibration from the rest of the gcode generation logic. This allows us to more easily call this method if we want to for other tools.

This commit also pushes code for obtaining the centre of mass from the simulation processor, for our new X, Y coordinates with the vibraiton tool. We now just need to call this generate_vibraiton method within the vacuum_pnp class